Support deletion of drafts (with confirmation)
This enables the 'Delete draft' button, which for now will delete a
document but in future should be changed to work on the current edition.

I added a new 'confirmation' flash type to support the interstitial
confirmation message, which will need a bit of frontend work.

One of the foreign keys (for lead images) was actually impeding a
document being deleted, as it had a race condition for the general
document<->image foreign key. I've changed the constraint to nullify
instead of cascade on delete, which breaks the race condition.