Ensure we discard drafts from the Publishing API
It's good to handle the error case when we implement the feature, but
I've gone one step further here and made 'error_deleting_draft' a new
publication state, to be consistent with 'error_sending_to_live', etc.
This means we can similarly show a 'Try again' button when the action
fails, but perhaps this is too much. It does seem weird that we show a
flash and a 'Try again' button in a separate panel.