Support deleting assets from asset manager
This adds #delete to the AssetManagerService so we can manage deleting
images and other assets, which will come in a subsequent PR.

I struggled quite a bit when adding this method, because there were a
few concepts that need to be understood. I've tried to make that easier:

  * asset_id: this was generated in AssetManagerService using 'url',
which came from AssetManagerFile, which came from Image. To make it
easier to understand how the value is generated, I've moved this method
directly onto the Image model, so that it's easier to see the data flow.

  * asset_manager_file_url: this is generated by Asset Manager itself.
Each test required 3 lines in order to mock it accurately, which was
distracting from the purpose of the test. I've DRY-ed this up into a
'in_asset_manager' trait on the Image factory.

  * AssetManagerFile: this was a nested class in the Image model, but
has no explicit relation to an Image and is only used by the
AssetManagerService. I've moved the class into AssetManagerService to
make this clearer and let the Image class concentrate on the cropping.