Add Papertrail for auditing
This adds the papertrail gem to the app and sets it up. Paperclip will
keep track of all of the changes in the models (Document, Image, User).

The reason for this is to keep an audit log of all of the changes.
While we're busy building this application, we're very likely to run
into a situation where a document has entered a particular state, and
we're not sure how this happened. Having a full log of changes will
help us a lot in debugging. And with time being linear, we won't be
able to add this feature retroactively if something goes wrong.

It's worth noting that this is a distinct & separate use case from the
history timeline
(https://github.com/alphagov/content-publisher/pull/270), which is a
feature designed for end-users. The two shouldn't be coupled.

Downside of having this:

- It slow all of the writes down. When a document is edited, there will
be multiple versions created because of the state changes.
- The versions table is likely to grow a lot. We'll need to make sure
that we clean up if the table becomes too large. There's a ticket on
the planning board for this: https://trello.com/c/XMrIPU88. This is
also the reason that we need to keep the history feature and this audit
trail separate, because we should be able to truncate the table without
losing end-user functionality.