Store role appointments on News and Press Releases
These are intended to replace ministers for the time being since there
is not a trivial way to get ministers nor does there seem to be any
special behaviour for ministers rather than other government
appointments.

The tricky thing though is what to name this label. I went for
Government appointment which is probably bad.