Add last_edited_by method to Document
This retrieves the user that last edited a document by querying the most recent
Timeline Entry with type `updated_content` or `updated_tags`. We have defined
"last edited" to mean a change to the document itself rather than a state
change such as sending for review or publishing.

We have decided to perform a query on Timeline Entries so that we're using
these entries as the source of truth for a document's history. We considered
associating a `last_edited_by` user to a document and updating this every time
a document is edited, which could be a simpler approach. However since we are
already storing the user in the Timeline Entry anyway it doesn't seem
particularly useful to store this information again elsewhere. This approach
also should also prevent any discrepancies that could occur between the user
stored on the Document and the Timeline Entries.