Display last edited by user on summary page
In the event that a document has not yet been edited, it makes sense to show
the name of the user that created it as that was the latest interaction with
the document.

In the event that a document has no creator, for example if the
document has been imported and the user is not present in this application,
then we display the same 'unknown' value as the 'Created by' field.