Add placeholder screens for retire/remove
This adds two separate placeholder screens for retiring/removing a
document, so that a user can see the feature isn't supported and know
what to do. I wanted to keep the new actions out of the already bulging
DocumentsController, so I've created two completely separate placeholder
routes, which will allow these features to be implemented separately.
In order to render the textual content on the page, I've used the same
pattern as for new_document/guidance.html.erb.