Simplify image_meta component and move choose link
The image_meta component is now less coupled from the actions it
displays, which are split into a single (optionally-tagged) 'primary'
action and a number of 'secondary' actions. This made it easy to move
the 'Choose image' link into the right place, and will make it easy to
add the upcoming 'Delete image' link as a secondary action.

Using the image IDs for the ID of the row means are specs are now more
robust, as they no longer depend on the ordering of the images on the
index page. This was also an opportunity to simplify the specs and add a
missing check to ensure we are sending the right data to the Publishing
API when we remove or choose an image.