Merge API error messages and simplify specs
This merges the API error messages for images so that they can be used
in the context of deleting an image - the previous messages were
specific to certain actions, like uploading an image. Fixing the specs
was an opportunity to make them a bit simpler by shortening some of the
steps and adding a missing step to check that setting a lead image works
even if the Publishing API is down.