Simplify upload spec and check for API request
This ensures we check the correct attributes are sent to the Publishing
API, as well checking as the appropriate TimelineEntry is created. A
redundant set of steps that published the document have been removed.
I've also removed the ambiguity when checking the caption/credit, which
were being set to the same text as the label for the field.