Handle Asset Manager down when cropping an image
The original spec focussed on cropping a lead image, but we support
cropping any image so I've renamed it accordingly and added additional
specs to check for the API down behaviour for the two APIs involved.

This also changes the crop endpoint so that an image_updated event is
properly created irrespective of whether the image is the lead image for
the document, which helps make this endpoint generic for all images.