Shift where upload spec checks for Pub API down
The previous spec that covered uploading an image with the Publishing
API down is now covered by the new crop spec. But the next stage in the
upload ('Save and choose') also required the Publishing API but was not
tested, so I've changed this spec to target this endpoint instead.