Replace 'next_screen' with 'wizard' tracking param
This fixes an issued where editing a non-lead image metadata created a
lead_image_updated event, while also changing the text of the button on
the metadata edit screen depending on which 'wizard' we are following.

Inverting the 'next_screen' behaviour and tracking that we're in a
'wizard' seems to make the redirect_to chain more obvious. It also seems
more straightforward to have separate endpoints for the update vs.
update-and-choose behaviours, as they differ in multiple ways, but also
because this clearly separates generic image behaviour from lead images.