Adds markdown toolbar to the editor
As a publisher
I need to add and format content with markdown
So that it appears on GOV.UK

We've iterated our markdown approach in Q2 and have prioritised the tool bar approach for private beta. We would like to test this approach in our first round  of research for Q2 next week.

The aim here is to make it easier for publishers to add markdown. We know from data and the survey that publishers spend time on finding the best markdown and editing the markdown. Also, for longer documents they spend lots of the time finding the errors in markdown.