Add image requirements for alt text / caption
This ensures we enter alt text for an image, as well as putting some
sensible limits on the maximum length of the text in these fields.

Although it's not possible to have multiple errors for each field in
this particular case, we'd like to adopt the pattern used here to
support it in other cases and maintain consistency.