Iterate publishing requirements to check presence
The previous publishing requirements are no longer applicable and we
want to iterate them to simply check for the presence of a summary and a
body. Instead of hard-coding the behaviour to the dynamic "body" field,
I've made it so that all dynamic field must be present on publish.

As all the requirements are now hard-coded, there's no need for the
'validations' hash for content fields. It's also good to remove this as
the new terminology of 'publishing requirements' and 'drafting
requirements' makes the lone term 'validations' pretty confusing.