Fix not checking for draft requirements on upload
Checking for draft requirements for images had only been applied when
editing the image. Attempting to add the requirements checking to the
lead images controller failed because "render 'document_images/edit"
looses track of the lead_image wizard param. So this commit pulls some
of the lead-image-specific functionality back into the document images
controller, which actually seems to make the flow a bit clearer.