Change topics service to emit a lookup table
Previously the topics service returned a nested hash ('tree') to
represent the taxonomy. This makes it difficult to lookup a particular
topic that's deeply nested. Using a lookup table as the data structure
makes it easy to lookup a topic, but can also be used to iterate over
the tree structure as long as we have an initial entry point into the
lookup table, which turns out to be the homepage (i.e. the root taxon).

Although level one topics are available as expanded links for the
root topic, these do not expand to deeper levels of the taxonony.
Therefore we need to make another API call for the expanded links of
each level one topic and merge these results into the expanded links we
got the for the root topic to produce an intermediate nested data
structure that we can easily iterate over to produce the final lookup
table.