Stub links for all documents to cope with topics
The summary page must now make a call to the Publishing API in order to
get the topics for a document. Unlike tags, we can't store these locally
and have to make the request, even if the document has no topics.

Adding this stub still left us with a few specs failing. This specs
reset WebMock in order to run detailed checks against the content of the
last request, but this also resets all the stubs that were registed. We
should instead reset just the request registry specifically - future
versions of WebMock should have a method to official support this.

https://github.com/bblimke/webmock/blob/d8686502442d9830dcccd24a1120ac08413d857a/lib/webmock/api.rb#L69