Enhance ability to detect live publications with imported whitehall documents
A document will have a live edition of some kind if it meets at least
two criteria:
(1) the publication state of the imported document record is 'published'
(2) it has at least two editions.

If it has a live edition, these flags must be set properly:
- publication_state must be set to 'sent-to-live'
- has_live_version_on_govuk must be set to true
- review state must be set to 'reviewed'

Previously the code was only checking for criterion (1).

Main changes:
- add document as parameter to methods that determine flag states and
  use the edition count method to support criterion (2)
- modify tests to accommodate criterion (2)
- modify most tests so they also check for has_live_version_on_govuk state
- change wording of test case descriptions because now we have three
  states to check which seem linked