Rescue exceptions in partials for documents/show
Previously we rescued API exceptions at the controller level for the
documents/show view, automatically rendering a 'show_api_down'
alternative. Duplicating the view seems to be a poor approach, as it's
not been kept up-to-date as the main 'show' view has evolved.

This removes the 'show_api_down' view and pushes the exception handling
into the 'tags' partial. Originally we avoided doing this as it means
the 'tags' partial has to have more knowledge about the work being done
by the partials it, in turn, renders. On reflection, it's not so bad ;-)