Make sure we timeout early when getting all topics
Creating the topics index requires looping over all of the L1 topics,
each of which can potentially take a long time to yield a response. If
the Publishing API is running slow, we accumulate a very long response
time at the application level, causing Nginx to time out.

This adds some superficial logic to monitor the total time it's taken to
accumulate the L1 topics (and the L0 topic) and abort if it's taking too
long. This means we can show an in-app error to the user, rather than a
Gateway Timeout at the Nginx level.