Enforce entering a change note when we publish
When the user sets the "major" update type for a live document, we
should ensure they enter a change note before they publish.

We currently auto-select the "major" update type each time a user
creates a new edition, so this updates the document factory to match.

Using interpolated values for locale keys leads to grammatical
problems (e.g. Enter a alt text) and we should avoid it.