Enable editing topics for a document
This adds a checkbox for each bullet point on the topics page and then
patches the document with the selected topics when the user clicks 'save'.

This was made easier by modifying the existing 'topics_for_document'
method to accept a document and return content_ids, rather than topic
hashes. Content IDs are required as the value for each checkbox, and it
was simple to change the other parts of the UI to work with Content IDs.