Adds attribute support to autocomplete to show contextual guidance
As a publisher
I need guidance on how what should I fill in in a form field
So I can complete my task

Currently the form elements jump when focussing due to the contextual guidance

The contextual guidance component doesn't work with Autocomplete (or any other component that replaces an input with another one). We need to update contextual guidance to support this type of component. An initial version was proposed but rejected in review as it was dealing with solely the Autocomplete (wasn't a generic solution).

Other updates on contextual guidance are hiding the guidance by default (avoid screen flashing) and overlap guidance over the next row (position absolute)

Ticket: https://trello.com/c/hpe5cNcL/349-update-contextual-guidance-component-to-work-with-progressively-enhanced-components