Cope when saving topics with the API down
This discards changes to the topics for a document and redirects to the
document summary in the event the Publishing API is down. Ideally we
would preserve the changes to allow the user to try again, but since we
also require the Publishing API to render the page, this is the easiest
thing to do for now.

This also tweaks some of the specs for topics and tags to make them more
consistent.