Add support for update_type and clean up test cases
Previous changes made it easier to identify imported documents as
having a live edition somewhere on GOV.UK.  In views/documents/show/_contents.html.erb,
the flag 'has_live_version_on_govuk' was used in a conditional to process a bit of
view code that required document.update_type to be defined.  This flag was not being
set in the importer, so when that bit of view code was being used it caused an error.

The importer now sets the update_type flag so that it may work correctly in that view.
So this flag change is a side effect of more imported documents being marked as having a
live version on GOV.UK

Changes to the tests for this feature have been modified to reflect this functionality
and to improve test documentation:
 - Added tests for update_type.  One for when the imported document has a minor change
   and another for the default case when it has a major one.
 - Used a context block to  make one of the 'lets' examples more closely tied to the
   one test case that uses it.  The block was moved to the end of the file.
 - Generalised comments to 'sets the correct states' rather than trying to list them all
   in one test comment.