Changing to reflect a clearer understanding of the distinction between 'publication_state' of a document and its 'is_live' flag. Changed name of private method 'is_live'to a better worded 'has_live_version_on_govuk'.
To make clear consider a document which has two editions, an early one whose
state is 'published' and a later one that has 'draft'.

'has_live_version_on_govuk': indicates whether a document has an edition which
is published. It would be true here because of the state of the first one.

'publication_state': is influenced entirely based on flag values of the most
current edition.  In this example, because the latest edition has a state of
draft, its publication state would be 'sent-to-draft'.