Separate creating a new edition from publishing
Previously the update_type and change_note were reset for a document
when it was published, such that the 'Create new edition button' had no
effect (it was just another link to the document edit page).

This changes the publishing service to not update the update_type or
change_note and instead move this logic into a new EditionsController,
which makes the change and then redirects to the edit page like before.