Replace and rename the document creation spec
The previous spec was primarily testing the change note and state
features, rather than the process of creating a document. The change
note aspect is now covered by the new edition spec.

Rather than re-write the spec to just focus on the state aspect, and in
slight conflict with the existing format specs, the new spec is a more
complete test of the workflow for a generic document.