Auto populate legacy topics (specialist sectors)
The new topic taxonomy has replaced a range of legacy alternatives,
except for 'topics' a.k.a specialist sectors. This makes sure we set
the specialist sectors for a document when the topics change, using the
legacy mappings for the selected topics and their parent lineage.

If the document already has specialist sectors, we will overwrite them.
This is a departure from the Whitehall behaviour, which we will validate
as with user testing. Hopefully we will retire specialist sectors before
the need arises to better support their existence ;-).