Check remote version when editing topics
Ww want to show an error message when trying to edit topics if the
remote version has changed. The Publishing API checks the previous
version of a document when trying to change it and returns a 409 if
trying to change a different version, so we just had to get the version
on the topics page and pass it to that method.