Push creating a new edition into drafting service
Previously the editions controller reset the update_type and change_note
for a document before redirecting the user to the content edit view. But
it's also possible for a user to manually navigate to edit views and
click save, and in this circumstance we should ensure this behaves the
same as explicitly creating a new edition. The document drafting service
already has a conditional for this purpose, so this simply re-uses that.