Add ContactsService to retrieve contacts
This is a deliberately bare bones service that is used to pull in
contact documents from the Publishing API.

My expectation is that this will need to be fleshed out further or in
different ways to meet the needs of creating the UI component to insert
contacts.

It does have a glaring issue that it would publish a draft contact if
someone is working on a draft of one at the time of usage. I'm hoping
this issue can be considered as part of the UI work.