Cleanup superfluous topics before patching links
Previously we assumed the frontend would ensure only leaf topics were
submitted on the edit page for topics; this assumption was false.

This changes the patch_links method to remove superfluous topics that
are an ancestor of another selected topic. To avoid implementing
multiple recursive methods, I extracted the recursive part of the
topic_breadcrumb method so that we can re-use it for ancestor cleanup.