Replace Contacts in Whitehall import process
This changes the body field on imported documents from Whitehall so that
the values for [Contact:123] are instead represented by the content_id.

It expects the JSON for a document to have a hash of contacts which
has a mapping of whitehall id => content id (e.g. { "123" =>
"12345678-1234-1234..." } which covers all the contacts in all the
translations and all the editions of a document.