Rename GovspeakService to GovspeakDocument
After chatting with Ben we agreed that this class now represents a
Govspeak document and doesn't really fit the name type of
GovspeakService. It still lives in the services directory as that seems
to be a bit of a lib directory - which is a problem to fix on another
day.