Refactor requirements so adding topics is easier
Extracting the existing publishing requirements into separate method
will make it easier to add an extra, complex requirement for topics.

Errors are still indexed by field, but the absence of errors is now
expressed as an empty array, which makes building the hash simpler with
the extracted checks, and also provides more consistency for the output.

This also adds an errors? method to abstract how we check the hash.