Refactor topics service into the document model
Previously all logic related to topics was contained within the
TopicsService. While this had the advantage of putting all the related
code in one place, the interface to topics (via a dumb hash) and the
combination of a document with its topics was convoluted.

As part of making topics a publishing requirement, we want to avoid
making a duplicate request to get the topics for a document. The
previous interface would require us to pass the topic service (or some
proxy) separately to the document as part of the setup.

This splits getting and setting the topics for a document into a new
class accessible by calling 'document.topics', which removes the need to
write setup for the service in multiple places. By convention, we expect
the result of 'document.topics' to return a collection of Topic objects.

Using an actual Topic model via a document simplifies the interface to
topics in the views. This also fixes a couple of small issues:

   - Making duplicate calls for expanded links when getting
level_one_topics
   - Changing the topic index code to refer to 'raw topics' instead of
'topics'