Ensure each document has at least one topic
Before a user publishes a document (or submits it for 2i), we should
attempt to show a warning if they have not associated the document with
at least one warning. When they try to publish (or submit), this warning
should become an error, or a fatal error if the Publishing API is down.