Load all contacts from Publishing API into Content Publisher
This allows us to present to users a list of all the contacts that are
available - and relies on Publishing API having the links part of the
GET editions endpoint: https://github.com/alphagov/publishing-api/pull/1383

It's a bit of a brute to load in all the contacts - at requests of 1000
a time it takes 4 HTTP requests, these seem to be pretty fast requests
but still on my VM I can hit 2.5 seconds for these so they aren't ideal.

I had some challenges with hashes indexed with symbols or with strings
(as the JSON responses return strings) and decided to stick with
everything being strings. If this feels inconsistent we can probably use
HashWithIndifferentAccess instead.