Use a common mechanism for checking requirements
Previously we had distinct classes to evaluate requirements on different
pages of the app. This was a bit misleading (e.g. 'DraftingRequirements'
was only intended for the document edit page) and assumed the user would
always follow the intended workflow such that the requirements were
enforced; this proved to be a false assumed (user aborts image wizard).

This merges the main requirements logic into a single, common mechanism
of 'checkers'. Putting everything in one place helpers keep things
consistent, but also enables us to re-use the same logic to show issues
on the document summary screen with customised messaging and links.