MVP implementation of inserting contacts to a document
This implements the design to insert contacts into a document that
involved loading a new page - with the intention of avoiding a modal
window and having an identical JS and No JS experience.

It involves allowing a user to submit the edit content with a different
button and because of that button changing the behaviour of the update
action to redirect the user to a select contact view. When this form is
submitted it then redirects the user back to the edit content form.

This approach is hopefully not something we will keep long term or use
again as it suffers from a number of issues:

- User can be prevented from getting to the insert contact screen by a
  validation error
- A new entry in the document history will be created each time the user
  submits either form
- We can't insert at a particular point in the textarea field easily and
  thus append to the end of the string
- We have to closely embed which field will be amended with the
  DocumentContactsController which doesn't make it particularly friendly
  to the format system

There are a number of approaches that could solve one or more of these
issues but will likely cause the complications and exceptions for this
to increase - whereas it probably makes sense to be entirely in the
concern of the govspeak editor.