Add an API down handler for adding contacts
This catches a GdsApi error and renders an API down page as per other
controllers and actions.

The test for this case ended up being a bit whacky because we want a
HTTP request to succeed and then redirect to a page where the Publishing
API is down. It's not too easy to set up the framework to test this
though as the redirect step is hidden. Hopefully my comment explains the
oddity a bit.