Add spec for drafting requirements on summary page
This feature only checks for the case of missing alt text, which is
currently an issue with the workflow. As there's now some repetition
between the features, it makes sense to move some of the tests out of the
image metadata and content edit features into rnit specs for the checkers.