Move testing requirements checkers into unit specs
Previously we tested all requirements logic through features, which is
difficult when a given field can have several requiremnts. It also lead
to some duplication, since we now show issues on multiple pages.

This moves the majority of the testing for requirements into unit tests
for the document and image checkers, as well as making the tests
exhaustive to cope with imported documents and changing requirements,
which existing or imported documents may not satisfy automatically.

Doing this also removes coupling between the features (e.g. topics).