Rename short/long_message to edit/summary_message
Previously we used short/long_message to allow for different messages to
be shown on the edit screen and the summary screen e.g. if there is an
issue with an image, on the summary screen we should indicate which one.

This renames 'short' and 'long' to 'edit' and 'summary' to better
indicate their use and avoid the slightly embarassing scenario where the
short message was longer than the long message ;-).