Add a service to retire a document
Build the backend service that withdraws ("retires" in the Content
Publisher terminology) from publishing-api.

Eventually this service will form the foundation for the full feature
that will allow users to retire content themselves, but for now it will
be used by a rake task that will be added in a later commit to allow
content to be retired manually.