Add a service to remove a document
Build the backend service that removes content from publishing-api.

Removed content no longer appears on GOV.UK

Eventually this service will form the foundation for the full feature
that will allow users to remove content themselves, but for now it will
be used by a rake task that will be added in a later commit to allow
content to be removed manually.