Add a rake task to retire documents manually
This task will have a few optional parameters, but the other unpublish
tasks will have more. The only parameter that will be required in all of
the tasks is `content_id`.

To make the experience consistent across all of the unpublish tasks, all
of the other parameters will be defined as environment variables.

Uses the climate_control gem to modify the value of the environment
variable on a per test basis.
This means that any modifications to the env vars won’t bleed into the
other tests.