Link to the unpublish form in the support app
Users need to know how to request for a document to be removed or
retired.

There is an existing form in the support app for users to request
Whitehall content is unpublished.

As the MVP we are just linking to this form. The plan is to either
update this form for use with Content Publisher or to create a brand
new form in future.