Split the “remove” rake task
Documents that are removed (but not redirected) and also have an optional explanatory note and alternative path.
Rather than try and incorporate all of these many options into one “Remove” task, it is simpler to separate them.
The optional notes and paths will be added in a later commit.

This is a follow on from splitting the “remove” service in an earlier commit.