Make support links link to production
As we will be allowing users to test this application in an integration
environment we want all support requests to still come from the
production environment.

It also might make sense to keep these production as we production
specific links to support in govspeak text as there's not a simple way
to provide an environment link there, so it wouldn't surprise me if
these remain linked to production indefinitely.