Add workflow to manually create a preview
In future we expect to have documents where the publication state is
'changes_not_sent_to_draft', as we stop relying on automatic preview.

This changes the actions shown when a document is in this state, so that
we show a 'Preview' button that can act as a last resort mechanism. From
the user's point of view, it behaves the same as the current 'Preview'
button, which takes them straight to the preview page for a document.