Fix specs that rely on 'in preview' actions
Previously we changed the actions for when a document is in the
'changes_not_sent_to_draft' state to not show the 'Publish' button.

A document created with the factory is actually in this state, so this
fixes the specs that need a document to have this action available.

This was also an opportunity to remove some duplication with the
'with_required_content_for_publishing' trait, now called 'publishable'.