Fail silently when automatically creating previews
Previously we would show an error message whenever a user changed some
part of the document and we (failed to) automatically create a preview.
The error was unrelated to the 'Save' action the user thought they were
performing, and now that we have a last resort mechanism to create a
preview, we no longer need to rely on auto preview (or show its errors).

The new, silent method of creating a preview is similar to invoking an
asynchronous job, where any failures would also be silent. I think this
is a precedent for removing the tests for when the Publishing API is
down, which reduces the boilerplate of writing new features and seems
like a lot of effort to test the effect of a single, declarative line.