Stop trying to create a preview if our checks fail
This caused a few tests to fail that check for the (silent) creation of
a preview. In the case of images, it made sense to rename the existing
'in_asset_manager' trait to 'in_preview' to reflect the common scenario
for when an image is in asset manager, after satisfying our requirements.