Fix incorrect error state if silent preview fails
Previously we changed the behaviour of saving changes to a document to
fail silently when creating a preview. Unfortunately this left the
document in the 'error_sending_to_draft' publication state. This fixes
the state to ensure its reset to 'changes_not_sent_to_draft' and
reinstates an API down spec to cover this scenario (lesson learnt ;-)).