Use /government/news prefix for our news formats
We want to avoid creating a new form of base path for news articles
on GOV.UK without further discussion, so for now we should stick to
the prefix used by Whitehall. This will generate an error if the base
path selected in Content Publisher happens to exist for a document
created in Whitehall. We will do subsequent work to detect this and give
better feedback to the user so they can fix the issue themself.