Only run the path checker on the edit page
This narrows the application of the path checker so that we only check
for a path conflict when saving the edit page. Previously we applied the
check every time we generated a preview, for which caching helped but it
still means we fire off arguably unnecessary requests every few minutes.

As we already coped with a path conflict by showing an error, it seems
more pragmatic to think about this check as a 'helper' and not worry too
much if it fails for now - most of the time we should correctly display
an issue at the right moment. If the coincidence of a base path conflict
and the Publishing API being down turns out to be higher than expected,
we can always go back to running this check more frequently on preview.