Fix markdown preview not working on Microsoft Edge
Previously we used Promise.prototype.finally to ensure we added a
particular CSS class irrespective of the outcome of the promise.
Unfortuantely this is only supported in the newest versions of MS Edge.

This moves and duplicates this one line into the .then and .catch blocks
as an alternative to using .finally. This is the only place we use this
method in the app, so it doesn't seem worth fixing with a polyfill.