Word-wrap SCSS helper applied to summary component
This borrows from this helpful CSS tricks article [1] a technique to
break up long lines of text that can break our applications layout.
Aside from extreme user testing (e.g. massive 100+ character words) the
most likely place we will see this is when a user has a URL in their
text.

This resolves the issue by forcing the word wrap. This does
unfortunately have a consequence that in IE11/Edge/Firefox we get breaks
in the middle of words rather than at the end (as they don't support the
non-standard property of break-word). I don't think there's a way around
that so we may have to decide to just drop this functionality if it
provides more confusion than problems it resolves.

[1]: https://css-tricks.com/snippets/css/prevent-long-urls-from-breaking-out-of-container/