Replace error logging with Sentry notifying
In these cases we shouldn't be expecting any HTTP errors they should
indicate either a problem with our calling or a problem with the service
itself.

Because of the configuration of govuk_app_config [1] timeout and
intermittent errors won't get reported to Sentry and instead will get
logged to statsd so that should save us from a mass of noisy
notifications. This will also mean that we don't miss genuine problems
that are only sent to the realms of Kibana.

[1]: https://github.com/alphagov/govuk_app_config/blob/47b90631399ef40d2fd74b8a715ad564b6a2ca54/lib/govuk_app_config/configure.rb#L14-L20